ATR-355: Refactoring.
