+++ src/Acuminator/Acuminator.Utilities/Acuminator.Utilities.csproj
@@ -129,7 +129,7 @@
     <Compile Include="Roslyn\Semantic\PXGraph\Infos\ProcessingDelegateInfo.cs" />
     <Compile Include="Roslyn\Semantic\PXGraph\ProcessingDelegatesWalker.cs" />
     <Compile Include="Roslyn\Semantic\Symbols\BqlDataTypeSymbols.cs" />
-    <Compile Include="Common\Constants.cs" />
+    <Compile Include="Roslyn\Constants.cs" />
     <Compile Include="Roslyn\Semantic\Symbols\EventSymbols.cs" />
     <Compile Include="Roslyn\Semantic\Symbols\BQLSymbols.cs" />
     <Compile Include="Roslyn\Semantic\Symbols\ExceptionSymbols.cs" />
