+++ src/Acuminator/Acuminator.Utilities/Roslyn/Semantic/Symbols/FieldAttributeSymbols.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.CodeAnalysis;
-using static Acuminator.Utilities.Roslyn.Semantic.Symbols.Constants;
+using static Acuminator.Utilities.Roslyn.Constants;
 
 namespace Acuminator.Utilities.Roslyn.Semantic.Symbols
 {
